refactor: Rename baseline component to surveys
