+++ lib/src/demographics/process_demographics_data.dart
@@ -4,7 +4,7 @@ import 'package:mdigit_span_tasks_ema/src/auth/participant.dart';
 import 'package:mdigit_span_tasks_ema/src/ema_data/ema_data.dart';
 import 'package:research_package/research_package.dart';
 
-import '../baseline/data/survey_data.dart';
+import '../surveys/data/survey_data.dart';
 
 /// Converts the [RPTaskResult] into a [SurveyData] object.
 void processDemographicsData(RPTaskResult results) {
