+++ lib/src/ema_data/ema_data.dart
@@ -1,6 +1,6 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:mdigit_span_tasks_ema/src/baseline/data/survey_data.dart';
-import 'package:mdigit_span_tasks_ema/src/baseline/data/survey_item_data.dart';
+import 'package:mdigit_span_tasks_ema/src/surveys/data/survey_data.dart';
+import 'package:mdigit_span_tasks_ema/src/surveys/data/survey_item_data.dart';
 
 class EMAData {
   final FirebaseFirestore db;
