+++ lib/src/routes.dart
@@ -4,7 +4,7 @@ import 'package:mdigit_span_tasks_ema/src/ema/ema_screen.dart';
 import 'package:mdigit_span_tasks_ema/src/task_list/view/task_list_page.dart';
 import 'package:mdigit_span_tasks_ema/src/ui_components/loading_screen.dart';
 
-import 'baseline/baseline_survey_screen.dart';
+import 'surveys/baseline_survey_screen.dart';
 import 'home/home.dart';
 import 'home/initial_route_middleware.dart';
