+++ test/src/surveys/data/survey_item_data_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:mdigit_span_tasks_ema/src/baseline/data/survey_item_data.dart';
+import 'package:mdigit_span_tasks_ema/src/surveys/data/survey_item_data.dart';
 import 'package:research_package/research_package.dart';
 
 void main() {
