+++ test/src/surveys/services/rp_converters_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:mdigit_span_tasks_ema/src/baseline/services/rp_converters.dart';
+import 'package:mdigit_span_tasks_ema/src/surveys/services/rp_converters.dart';
 import 'package:research_package/research_package.dart';
 
 void main() {
